+++ code/modules/client/preferences/middleware/food.dm
@@ -3,18 +3,21 @@
 #define MINIMUM_REQUIRED_DISLIKES 3
 #define MAXIMUM_LIKES 3
 
-// Performance and RAM friendly menu. You can thank me later.
-GLOBAL_DATUM_INIT(food_prefs_menu, /datum/food_prefs_menu, new)
-
 // Hahahaha, it LIVES!
+/datum/preference_middleware/food
+	action_delegations = list(
+		"reset_food" = PROC_REF(reset_food),
+		"toggle_food" = PROC_REF(toggle_food),
+		"change_food" = PROC_REF(change_food),
+	)
 
 /datum/preference_middleware/food/apply_to_human(mob/living/carbon/human/target, datum/preferences/preferences)
 	if(!length(preferences.food_preferences) || isdummy(target))
 		return
 
-	var/fail_reason = GLOB.food_prefs_menu.is_food_invalid(preferences)
+	var/fail_reason = is_food_invalid(preferences)
 	if(fail_reason)
-		to_chat(preferences.parent, span_announce("Your food preferences can't be set because of [fail_reason] choices! Please check your preferences!")) // Sorry, but I don't want folk sleeping on this.
+		to_chat(preferences.parent, span_announce("Your food preferences can't be set! Reason: \"[fail_reason]\" Please check your preferences!")) // Sorry, but I don't want folk sleeping on this.
 		return
 
 	var/datum/species/species = target.dna.species
@@ -34,79 +37,58 @@ GLOBAL_DATUM_INIT(food_prefs_menu, /datum/food_prefs_menu, new)
 			if(FOOD_PREFERENCE_TOXIC)
 				species.toxic_food |= GLOB.food_ic_flag_to_bitflag[food_entry]
 
-/// Food prefs menu datum. Global datum for performance and memory reasons.
-/datum/food_prefs_menu/ui_interact(mob/dead/new_player/user, datum/tgui/ui)
-	ui = SStgui.try_update_ui(user, src, ui)
-	if(!ui)
-		ui = new(user, src, "FoodPreferences", "Food Preferences")
-		ui.set_autoupdate(FALSE)
-		ui.open()
+/datum/preference_middleware/food/proc/reset_food(list/params, mob/user)
+	qdel(preferences.food_preferences)
+	preferences.food_preferences = list()
+	return TRUE
 
-/datum/food_prefs_menu/ui_act(action, list/params, datum/tgui/ui, datum/ui_state/state)
-	. = ..()
+/datum/preference_middleware/food/proc/toggle_food(list/params, mob/user)
+	if(!preferences.food_preferences)
+		preferences.food_preferences = list()
+	preferences.food_preferences["enabled"] = !preferences.food_preferences["enabled"]
+	return TRUE
 
-	var/datum/preferences/preferences = ui?.user?.client?.prefs
+/datum/preference_middleware/food/proc/change_food(list/params, mob/user)
+	var/food_name = params["food_name"]
+	var/food_preference = params["food_preference"]
 
-	switch(action)
-		if("reset")
-			qdel(preferences.food_preferences)
-			preferences.food_preferences = list()
-			return TRUE
+	if(!food_name || !preferences || !food_preference || !(food_preference in list(FOOD_PREFERENCE_LIKED, FOOD_PREFERENCE_NEUTRAL, FOOD_PREFERENCE_DISLIKED, FOOD_PREFERENCE_TOXIC)))
+		return TRUE
 
-		if("toggle")
-			preferences.food_preferences["enabled"] = !preferences.food_preferences["enabled"]
-			return TRUE
+	// Do some simple validation for max liked foods. Full validation is done on spawn and in the actual menu.
+	var/liked_food_length = 0
 
-		if("change_food")
+	for(var/food_entry in preferences.food_preferences)
+		var/list/food_points_entry = GLOB.food_ic_flag_to_point_values[food_entry]
+		if(length(food_points_entry) >= FOOD_PREFERENCE_OBSCURE && food_points_entry[FOOD_PREFERENCE_OBSCURE])
+			continue
 
-			var/food_name = params["food_name"]
-			var/food_preference = params["food_preference"]
+		if(preferences.food_preferences[food_entry] == FOOD_PREFERENCE_LIKED)
+			liked_food_length++
+			if(liked_food_length > MAXIMUM_LIKES)
+				preferences.food_preferences.Remove(food_entry)
 
-			if(!food_name || !preferences || !food_preference || !(food_preference in list(FOOD_PREFERENCE_LIKED, FOOD_PREFERENCE_NEUTRAL, FOOD_PREFERENCE_DISLIKED, FOOD_PREFERENCE_TOXIC)))
-				return TRUE
+	if(liked_food_length > MAXIMUM_LIKES || (food_preference == FOOD_PREFERENCE_LIKED && liked_food_length == MAXIMUM_LIKES)) // Equals as well, if we're setting a liked food!
+		return TRUE // Fuck you, look your mistakes in the eye.
 
-			// Do some simple validation for max liked foods. Full validation is done on spawn and in the actual menu.
-			var/liked_food_length = 0
+	preferences.food_preferences[food_name] = food_preference
+	return TRUE
 
-			for(var/food_entry in preferences.food_preferences)
-				var/list/food_points_entry = GLOB.food_ic_flag_to_point_values[food_entry]
-				if(length(food_points_entry) >= FOOD_PREFERENCE_OBSCURE && food_points_entry[FOOD_PREFERENCE_OBSCURE])
-					continue
-
-				if(preferences.food_preferences[food_entry] == FOOD_PREFERENCE_LIKED)
-					liked_food_length++
-					if(liked_food_length > MAXIMUM_LIKES)
-						preferences.food_preferences.Remove(food_entry)
-
-			if(liked_food_length > MAXIMUM_LIKES || (food_preference == FOOD_PREFERENCE_LIKED && liked_food_length == MAXIMUM_LIKES)) // Equals as well, if we're setting a liked food!
-				return TRUE // Fuck you, look your mistakes in the eye.
-
-			preferences.food_preferences[food_name] = food_preference
-			return TRUE
-
-/datum/preferences/ui_state(mob/user)
-	return GLOB.always_state
-
-/datum/food_prefs_menu/ui_status(mob/user, datum/ui_state/state)
-	return UI_INTERACTIVE // Prefs can be accessed from anywhere.
-
-/datum/food_prefs_menu/ui_static_data(mob/user)
+/datum/preference_middleware/food/get_ui_static_data(mob/user)
 	return list(
 		"food_types" = GLOB.food_ic_flag_to_point_values,
 	)
 
-/datum/food_prefs_menu/ui_data(mob/user)
-	var/datum/preferences/preferences = user.client.prefs
-
+/datum/preference_middleware/food/get_ui_data(mob/user)
 	return list(
-		"selection" = preferences.food_preferences,
-		"points" = calculate_points(preferences),
-		"enabled" = preferences.food_preferences["enabled"],
-		"invalid" = is_food_invalid(preferences),
+		"food_selection" = preferences.food_preferences,
+		"food_points" = calculate_points(preferences),
+		"food_enabled" = preferences.food_preferences["enabled"],
+		"food_invalid" = is_food_invalid(preferences),
 	)
 
 /// Checks the provided preferences datum to make sure the food pref values are valid. Does not check if the food preferences value is null.
-/datum/food_prefs_menu/proc/is_food_invalid(datum/preferences/preferences)
+/datum/preference_middleware/food/proc/is_food_invalid(datum/preferences/preferences)
 	var/liked_food_length = 0
 	var/disliked_food_length = 0
 	var/toxic_food_length = 0
@@ -127,16 +109,16 @@ GLOBAL_DATUM_INIT(food_prefs_menu, /datum/food_prefs_menu, new)
 				toxic_food_length++
 
 	if(liked_food_length > MAXIMUM_LIKES)
-		return "too many like choices"
+		return "Too many like choices."
 	if(disliked_food_length < MINIMUM_REQUIRED_DISLIKES)
-		return "too few dislike choices"
+		return "Too few dislike choices."
 	if(toxic_food_length < MINIMUM_REQUIRED_TOXICS)
-		return "too few toxic choices"
+		return "Too few toxic choices."
 	if(calculate_points(preferences) < 0)
-		return "not enough points"
+		return "Not enough points."
 
 /// Calculates the deviance points for food.
-/datum/food_prefs_menu/proc/calculate_points(datum/preferences/preferences)
+/datum/preference_middleware/food/proc/calculate_points(datum/preferences/preferences)
 	var/points = 0
 
 	for(var/food_entry in preferences.food_preferences)
