+++ tgui/packages/tgui/interfaces/PreferencesMenu/data.ts
@@ -184,6 +184,13 @@ export type PreferencesMenuData = {
   name_to_use: string;
 
   window: Window;
+
+  food_selection: Record<string, number>;
+  food_points: number;
+  food_enabled: BooleanLike;
+  food_invalid: string;
+
+  food_types: Record<string, Record<string, number>>;
 };
 
 export type ServerData = {
